requirement.txt
pyjwt
flask
python-dotenv
flask_mysqldb